celery_voice/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
cartesia==1.0.13
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.7
deepgram-sdk==3.7.3
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
iterators==0.2.0
jiter==0.5.0
kombu==5.4.2
logger==1.4
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
openai==0.28.0
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.1
tqdm==4.66.5
twilio==9.3.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
yarl==1.13.1
